+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   - Transaction date is now 'Journaaldatum' and not 'Boekingsdatum' for ASN CSV.
   - Enhanced documentation for ASN CSV.
+  - The transaction id for ASN CSV is now <'Journaaldatum' in yyyymmdd format>.<'Volgnummer transactie'>
 
 ## [1.4.0] - 2022-07-19